Fixed crash in esp_tls_conn_write.
esp_tls_conn_write is implemented as an inline function which
calls the write() function pointer in the esp_tls_t structure.
If mbedtls/esp_config.h is not included then CONFIG_MBEDTLS_SSL_PROTO_DTLS
does not get defined, and in turn MBEDTLS_SSL_DTLS_BADMAC_LIMIT
gets defined. This results in the structure having a different
size in esp_tls.c (which allocates and initializes it), and when
used outside of ESP-IDF.